Remove torch and ax-platform from normal dependencies. axPlatofrm has been moved to poetry.workflow
