+++ pyproject.toml
@@ -14,10 +14,8 @@ repository = "https://github.com/AntoineCollin/scMusketeers/"
 python = "^3.10"
 poetry = "^1.8.1"
 numpy = "^1.23"
-torch = "^2.3.1"
 keras = "^3.3.3"
 scanpy = "^1.10.1"
-ax-platform = "^0.4.0"
 neptune = "^1.10.4"
 pympler = "^1.0.1"
 tensorflow = "^2.10.0"
@@ -48,6 +46,14 @@ twine = "^5.1.1"
 scbalance = "^1.2.0"
 celltypist = "^1.6.3"
 scmappy = "^0.2"
+scvi = "^0.6.8"
+
+
+[tool.poetry.group.optim.dependencies]
+scvi = "^0.6.8"
+
+
+[tool.poetry.group.workflow.dependencies]
 scvi = "^0.6.8"
 
 [build-system]
