+++ scmusketeers/__init__.py
@@ -1,1 +1,14 @@
-from . import arguments, tools, transfer
+print("Import all modules")
+print("import arguments")
+from . import arguments
+print("import tools")
+from . import tools
+print("import transfer")
+from . import transfer
+# print("import workflow")
+# from . import workflow
+print("__main__")
+from . import __main__
+print("run_musk")
+from . import run_musk
+print("Finished")
